+++ packages/mcp-server/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.35.0 ((1/5/2026, 10:03 AM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.34.2 ((1/2/2026, 09:41 AM PST))
 
 This is an artificial version bump with no new change.